requirements.txt
Django==4.1.5
django_ajax_datatable==4.4.5